requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
idna==2.10
lxml==4.5.2
pycparser==2.20
pytelegram-bot==0.1
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.25.10
xlwt==1.3.0
